3.1.1_update

+++ src/main/java/web/service/UserService.java
@@ -5,8 +5,8 @@ import web.model.User;
 import java.util.List;
 
 public interface UserService {
-    public User getUser(int id);
-    void deleteUser(int id);
+    public User getUser(long id);
+    void deleteUser(long id);
     void addUser(User user);
     void updateUser(User user);
     List<User> listUsers();
